@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 
 import { SharedModule } from '../../shared/shared.module';
+import { AuthenticationService } from '../authentication/authentication.service';
+import { Roles } from '../authentication/roles';
 
 @Component({
   selector: 'app-home',
@@ -9,9 +11,13 @@ import { SharedModule } from '../../shared/shared.module';
 })
 export class HomeComponent implements OnInit {
 
-  constructor() { }
+  constructor(private authService: AuthenticationService) { }
 
   ngOnInit() {
   }
 
+  hasAdminRole() {
+    return this.authService.decodedToken().role === Roles.admin ||
+     this.authService.decodedToken().role === Roles.sebeka_gubae;
+  }
 }
